# FIX: Update hash tables of affected profiler instances when rewriting code objects 