Closes #350. Includes #349, which by extension closes #348. New changes are from 7338799 onwards.

Code changes
----
- `line_profiler/_line_profiler.pyx::LineProfiler`
  - `.add_function()`
    - Now synchronizing the hash tables of instances profiling the same function object whenever it replaces its code object
    - No longer creating separate `.dupes_map` items for padded bytecodes
  - `._all_paddings`, `._all_instances_by_funcs`  
    New private class attributes for use by `.add_function()`

Test changes
----
- `tests/test_line_profiler.py`
  - `test_{{class,static,bound,partial}method,partial,[cached_]property}_decorator()`, `test_profile_generated_code()`  
    Minor refactoring with a utility function for more readable output
  - `test_multiple_profilers_identical_bytecode()`
    New test checking that the order in which profilers are used to decorate functions shouldn't affect the output (up to overhead)
  - `test_aggregate_profiling_data_between_code_versions()`  
    New test checking that existing profiling data should persist and be aggregated with new data after another profiler replaced the code object of a profiled function

Performance impact
----
Test suite run 5 times with the command `pytest 'not (test_duplicate or test_aggregate or identical_bytecode)'`, i.e. excluding the tests added by this PR (and #349); impact on performance should be minimal.

| Setup\Code (mean, best (s)) | `main` | `multi-desync-fix` (this PR) |
| :-: | :-: | :-: |
| Python 3.8.20, pytest 8.3.5 | 7.605, 7.518 | 7.523, 7.426 |
| Python 3.13.3, pytest 8.4.0 | 9.835, 9.589 | 9.777, 9.662 |