@TTsangSC 2025-07-08
Just finished rebasing. @Erotemic, maybe you'll find it more convenient to look at this PR (which contains #349) to resolve both bugs at once? (Or maybe not – up to you of course.) Cheers.