@Erotemic 2025-07-09
I think this looks good in general. I'll let you make any cleanups, and tell me when you are ready to merge.